requirements.txt
mysql-python==1.2.5
pyyaml==3.11.0
numpy==1.14.0
pandas==0.22.0
scipy==0.14.1
scikit-learn==0.19.1
matplotlib==1.4.2
mlencoders
